feat: Remove hard dependency on styled-components
BREAKING CHANGE: Switched to supporting React.forwardRef and object styles by default, so the module is more generic and doesn’t rely on styled-components. Can be used with emotion as well. The reason for this change is that now styled-components supports object styles.

This will probably not work well with styled-components until its TypeScript typings are updated to support object styles.